feat(userland): glibc-ABI compatibility for the JVM stack and a Minecraft launcher
The prebuilt JDK and its native libraries are glibc-ABI objects, so they import
symbols MakaOS's libc did not define and expect layouts it did not provide.
This fills those gaps and adds the launcher used to drive Minecraft on top.

libc additions:

  * math_c99.c - the C99 doubles, the long-double (l) variants and the float (f)
    variants that the host libstdc++ <cmath> and <cstdlib> expect from a
    complete libc. The existing math.c already had the core double/float set.
  * cxxabi_compat.c - C++-mangled aliases for C functions that MakaOS's
    libstdc++.a imports. That libstdc++ was built against sysroot headers which
    did not wrap these in extern "C", so its objects reference the mangled names
    (for example _Z4statPKcP4stat for stat).
  * glibc_compat.c - the _FORTIFY_SOURCE __*_chk entry points that prebuilt
    .so's compiled with -D_FORTIFY_SOURCE reference, such as __snprintf_chk from
    jdk.management's libmanagement_ext.so.
  * libc.c - define __environ alongside environ. glibc exposes the array under
    __environ with environ as the weak alias, and prebuilt .so's import the
    former.
  * unistd.c - multithreaded-fork safety: fork copies only the calling thread
    but the whole address space, including any allocator lock a sibling thread
    happened to hold.
  * pthread.c - larger default thread stack for glibc-ABI software.
  * headers - fenv.h, math.h, stdint.h, stdlib.h, ucontext.h, sys/syscall.h plus
    new features.h, sys/procfs.h, sys/statfs.h and sys/swap.h.
  * sdl_port_stubs.c - additional stubs for the SDL port.

Build and tooling:

  * build.sh compiles and links math_c99.c, and under MCTEST stages the compiled
    McLaunch classes into the image at /mc/makaos.
  * userland/apps/mclaunch/McLaunch.java - the launcher entry point.
  * scripts/fetch-minecraft.sh stages the client and its libraries;
    scripts/boot.sh and scripts/run-quake-shot.sh are boot/capture helpers;
    scripts/run-gl.sh updated alongside them.

+++ userland/libc/libc.h
@@ -34,11 +34,11 @@ int* __errno_location(void);
 typedef unsigned char      uint8_t;
 typedef unsigned short     uint16_t;
 typedef unsigned int       uint32_t;
-typedef unsigned long long uint64_t;
+typedef unsigned long      uint64_t;   // LP64: 64-bit == long, matching glibc
 typedef signed char        int8_t;
 typedef signed short       int16_t;
 typedef signed int         int32_t;
-typedef signed long long   int64_t;
+typedef signed long        int64_t;    // LP64: 64-bit == long, matching glibc
 typedef uint64_t           size_t;
 typedef int64_t            ssize_t;
 typedef int64_t            intmax_t;
